+++ tests/unit/components/pokemon/PokemonItemCharacteristics.test.js
@@ -11,6 +11,8 @@ describe('LogoAndBanner', () => {
         weight: 80,
         height: 17,
         color: 'yellow',
+        shape: 'upright',
+        generation: 'i',
       },
     });
   });
@@ -25,18 +27,28 @@ describe('LogoAndBanner', () => {
   });
 
   it('displays the correct weight', () => {
-    const weight = wrapper.find('.status span:nth-child(2)');
+    const weight = wrapper.findAll('span').at(1);
     expect(weight.text()).toBe('80 pounds');
   });
 
   it('displays the correct height', () => {
-    const height = wrapper.find('.status span:nth-child(4)');
+    const height = wrapper.findAll('span').at(3);
     expect(height.text()).toBe('17"');
   });
 
   it('displays the correct color', () => {
-    const color = wrapper.find('.status span:nth-child(6)');
+    const color = wrapper.findAll('span').at(5);
     expect(color.text()).toBe('yellow');
     expect(color.attributes('style')).toContain('color: yellow');
+  });
+
+  it('displays the correct shape', () => {
+    const color = wrapper.findAll('span').at(7);
+    expect(color.text()).toBe('upright');
+  });
+
+  it('displays the correct generation', () => {
+    const color = wrapper.findAll('span').at(9);
+    expect(color.text()).toBe('I');
   });
 });